chore: delete test_pysqlx_vs_psycopg2.py
